+++ core/BaseQuery.js
@@ -1,6 +1,7 @@
 const HError = require('./HError')
 const Query = require('./Query')
 const utils = require('./utils')
+const constants = require('./constants')
 
 class BaseQuery {
   constructor() {
@@ -9,7 +10,7 @@ class BaseQuery {
 
   _initQueryParams() {
     this._queryObject = {}
-    this._limit = 20
+    this._limit = null
     this._offset = 0
     this._orderBy = null
     this._keys = null
@@ -70,7 +71,7 @@ class BaseQuery {
 
   _handleAllQueryConditions() {
     let conditions = {}
-    conditions.limit = this._limit
+    conditions.limit = this._limit === null ? constants.QUERY_LIMITATION_DEFAULT : this._limit
     conditions.offset = this._offset
 
     if (this._orderBy) {